correction prévision immunité, remplacement de l'ancienne classe méthodes des moindres carrés par régression linéaire

+++ src/main/java/com/pantheonsorbonne/infocovid/domain/model/PrevisionImmunite.java
@@ -34,7 +34,7 @@ public class PrevisionImmunite {
 	     *
 	     */
 	    @JsonProperty("nbVaccinQuotidien")
-	    private ArrayList<Integer> nbVaccinQuotidien;
+	    private ArrayList<Double> nbVaccinQuotidien;
 	
 
 }
